+++ cryptofeed/exchanges/binance_futures.py
@@ -12,7 +12,7 @@ from typing import List, Tuple, Callable, Dict
 from yapic import json
 
 from cryptofeed.connection import AsyncConnection, HTTPPoll, HTTPConcurrentPoll
-from cryptofeed.defines import BINANCE_FUTURES, FUNDING, LIQUIDATIONS, OPEN_INTEREST
+from cryptofeed.defines import BALANCES, BINANCE_FUTURES, FUNDING, LIQUIDATIONS, OPEN_INTEREST, POSITIONS
 from cryptofeed.exchanges.binance import Binance
 from cryptofeed.exchanges.mixins.binance_rest import BinanceFuturesRestMixin
 from cryptofeed.types import OpenInterest
@@ -23,13 +23,15 @@ LOG = logging.getLogger('feedhandler')
 class BinanceFutures(Binance, BinanceFuturesRestMixin):
     id = BINANCE_FUTURES
     symbol_endpoint = 'https://fapi.binance.com/fapi/v1/exchangeInfo'
+    listen_key_endpoint = 'listenKey'
     valid_depths = [5, 10, 20, 50, 100, 500, 1000]
     valid_depth_intervals = {'100ms', '250ms', '500ms'}
     websocket_channels = {
         **Binance.websocket_channels,
         FUNDING: 'markPrice',
         OPEN_INTEREST: 'open_interest',
-        LIQUIDATIONS: 'forceOrder'
+        LIQUIDATIONS: 'forceOrder',
+        POSITIONS: POSITIONS
     }
 
     @classmethod
@@ -103,6 +105,81 @@ class BinanceFutures(Binance, BinanceFuturesRestMixin):
                 ret.append((PollCls(addrs, self.id, delay=60.0, sleep=self.open_interest_interval, proxy=self.http_proxy), self.subscribe, self.message_handler, self.authenticate))
         return ret
 
+    async def _account_update(self, msg: dict, timestamp: float):
+        """
+        {
+        "e": "ACCOUNT_UPDATE",                // Event Type
+        "E": 1564745798939,                   // Event Time
+        "T": 1564745798938 ,                  // Transaction
+        "a":                                  // Update Data
+            {
+            "m":"ORDER",                      // Event reason type
+            "B":[                             // Balances
+                {
+                "a":"USDT",                   // Asset
+                "wb":"122624.12345678",       // Wallet Balance
+                "cw":"100.12345678",          // Cross Wallet Balance
+                "bc":"50.12345678"            // Balance Change except PnL and Commission
+                },
+                {
+                "a":"BUSD",
+                "wb":"1.00000000",
+                "cw":"0.00000000",
+                "bc":"-49.12345678"
+                }
+            ],
+            "P":[
+                {
+                "s":"BTCUSDT",            // Symbol
+                "pa":"0",                 // Position Amount
+                "ep":"0.00000",            // Entry Price
+                "cr":"200",               // (Pre-fee) Accumulated Realized
+                "up":"0",                     // Unrealized PnL
+                "mt":"isolated",              // Margin Type
+                "iw":"0.00000000",            // Isolated Wallet (if isolated position)
+                "ps":"BOTH"                   // Position Side
+                }，
+                {
+                    "s":"BTCUSDT",
+                    "pa":"20",
+                    "ep":"6563.66500",
+                    "cr":"0",
+                    "up":"2850.21200",
+                    "mt":"isolated",
+                    "iw":"13200.70726908",
+                    "ps":"LONG"
+                },
+                {
+                    "s":"BTCUSDT",
+                    "pa":"-10",
+                    "ep":"6563.86000",
+                    "cr":"-45.04000000",
+                    "up":"-1423.15600",
+                    "mt":"isolated",
+                    "iw":"6570.42511771",
+                    "ps":"SHORT"
+                }
+            ]
+            }
+        }
+        """
+        for balance in msg['a']['B']:
+            await self.callback(BALANCES,
+                                feed=self.id,
+                                symbol=balance['a'],
+                                timestamp=self.timestamp_normalize(msg['E']),
+                                receipt_timestamp=timestamp,
+                                wallet_balance=Decimal(balance['wb']))
+        for position in msg['a']['P']:
+            await self.callback(POSITIONS,
+                                feed=self.id,
+                                symbol=self.exchange_symbol_to_std_symbol(position['s']),
+                                timestamp=self.timestamp_normalize(msg['E']),
+                                receipt_timestamp=timestamp,
+                                position_amount=Decimal(position['pa']),
+                                entry_price=Decimal(position['ep']),
+                                unrealised_pnl=Decimal(position['up']))
+
     async def message_handler(self, msg: str, conn: AsyncConnection, timestamp: float):
         msg = json.loads(msg, parse_float=Decimal)
 
@@ -112,6 +189,13 @@ class BinanceFutures(Binance, BinanceFuturesRestMixin):
             if self.concurrent_http:
                 return create_task(coro)
             return await coro
+
+        # Handle account updates from User Data Stream
+        if self.requires_authentication:
+            msg_type = msg.get('e')
+            if msg_type == 'ACCOUNT_UPDATE':
+                await self._account_update(msg, timestamp)
+            return
 
         # Combined stream events are wrapped as follows: {"stream":"<streamName>","data":<rawPayload>}
         # streamName is of format <symbol>@<channel>