+++ CHANGES.md
@@ -32,6 +32,7 @@
  * Feature: Remove pandas dependency
  * Breaking Change: Rewrite all rest endpoints to support sync and async versions of the endpoint.
  * Feature: Add dYdX REST endpoints
+ * Feature: Add Binance authentication for User Data Streams
  * Feature: Add support for Binance trading REST API
  * Bugfix: Fix typo by renaming rest_options to order_options
  * Bugfix: Use correct max depth for Binance (and its child classes).